refactor autotrap
